Extract out cast hit and cast result into separate files
