+++ Assets/Code/_Common/Collisions/CastHit.cs
@@ -1,0 +1,21 @@
+using UnityEngine;
+
+
+namespace PQ.Common.Collisions
+{
+    public struct CastHit
+    {
+        public readonly Vector2 point;
+        public readonly Vector2 normal;
+        public readonly float distance;
+        public readonly Collider2D collider;
+
+        public CastHit(Vector2 point, Vector2 normal, float distance, Collider2D collider)
+        {
+            this.point = point;
+            this.normal = normal;
+            this.distance = distance;
+            this.collider = collider;
+        }
+    }
+}
