+++ Assets/Code/_Common/Collisions/CastHit.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cd62df07cffefde47afb22ae7446e1fd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
