+++ Assets/Code/_Common/Collisions/CastResult.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9a6d8f81f4c44d642a03ac27b873b42a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
